Implement factory method for employent type

+++ Sprout.Exam.WebApp/Sprout.Exam.Business/DataTransferObjects/EmployeePayrollDto.cs
@@ -3,6 +3,6 @@
     public class EmployeePayrollDto : BaseSaveEmployeeDto
     {
         public int Id { get; set; }
-        public int SalaryNetPay { get; set; }
+        public decimal SalaryNetPay { get; set; }
     }
 }
